This repo is meant to help me explore what...
Files.App.Utils.Cloud.GoogleDriveCloudDetector.GetProviders()

currently does, and what we want it to do.

In Files.App.Utils.Cloud.GoogleDriveCloudDetector.GetProviders(), first we get a
Sqlite database file at the following path:

"$($env:localappdata)/Google/DriveFS/root_preference_sqlite.db"

Then we copy that database to a temporary folder.

Then we run the following SQL queries against the copy:

SELECT * FROM roots

SELECT * FROM media WHERE fs_type=10

Then we traverse the ITEMs in each of the two response sets and create a new
GoogleDrive-type CloudProvider for each ITEM, with the CloudProvider.SyncFolder
property set to a (formatted) part of each ITEM's contents.

So far, in this scratch repo, I have confirmed that when Google Drive is mounted
as a folder, this Sqlite method for getting the mount paths DOESN'T work.

"When Google Drive is mounted as a folder" means when you have the "Google Drive
streaming location" setting set to "Folder" in Google Drive Preferences.

I have also confirmed that when Google Drive is mounted as a folder, the
registry method for getting the mount paths

( described here:
https://github.com/files-community/Files/issues/15698#issuecomment-2195215312 )

DOES work.

Next I will test the behavior of each method when Google Drive is mounted as a
dedicated drive.